refactor: Added applied filters to mobile filter modal
