+++ features/collection/components/FilterableProductSection.tsx
@@ -1,39 +1,22 @@
-import {
-   Button,
-   Collapse,
-   HStack,
-   Tag,
-   TagCloseButton,
-   TagLabel,
-   VStack,
-   Wrap,
-   WrapItem,
-} from '@chakra-ui/react';
+import { HStack, VStack } from '@chakra-ui/react';
 import { Card } from '@components/Card';
 import {
    CollectionFilters,
    CollectionPagination,
    CollectionProducts,
-   Toolbar,
-   ProductViewType,
-   NumberOfHits,
    FiltersModal,
+   NumberOfHits,
    OpenFiltersButton,
-   SortBySelect,
-   SearchInput,
-   ProductViewSwitch,
-   ProductViewListButton,
    ProductViewGridButton,
+   ProductViewListButton,
+   ProductViewSwitch,
+   ProductViewType,
+   SearchInput,
+   SortBySelect,
+   Toolbar,
 } from '@features/collection/components';
-import { formatFacetName } from '@features/collection/utils';
-import {
-   AtomicFilter,
-   useAtomicFilters,
-   useClearFilter,
-   useSearchStateContext,
-} from '@lib/algolia';
-import { assertNever } from '@lib/utils';
 import * as React from 'react';
+import { AppliedFilters } from './AppliedFilters';
 
 export const FilterableProductSection = React.memo(() => {
    const [productViewType, setProductViewType] = React.useState(
@@ -88,7 +71,7 @@ export const FilterableProductSection = React.memo(() => {
                <CollectionFilters />
             </FilterCard>
             <VStack align="stretch" flex={1}>
-               <AppliedFilterSection />
+               <AppliedFilters />
                <Card
                   flex={1}
                   alignItems="center"
@@ -116,74 +99,5 @@ const FilterCard = ({ children }: React.PropsWithChildren<unknown>) => {
       >
          {children}
       </Card>
-   );
-};
-
-const AppliedFilterSection = () => {
-   const atomicFilters = useAtomicFilters();
-   const clear = useClearFilter();
-   const clearAllFilters = React.useCallback(() => {
-      clear();
-   }, [clear]);
-   return (
-      <Collapse in={atomicFilters.length > 0} animateOpacity unmountOnExit>
-         <Wrap w="full" align="center">
-            {atomicFilters.map((filter) => {
-               return (
-                  <WrapItem key={filter.id}>
-                     <FilterTag filter={filter} />
-                  </WrapItem>
-               );
-            })}
-            <WrapItem>
-               <Button
-                  variant="link"
-                  size="sm"
-                  colorScheme="brand"
-                  onClick={clearAllFilters}
-               >
-                  Clear All
-               </Button>
-            </WrapItem>
-         </Wrap>
-      </Collapse>
-   );
-};
-
-interface FilterTagProps {
-   filter: AtomicFilter;
-}
-
-const FilterTag = ({ filter }: FilterTagProps) => {
-   const state = useSearchStateContext();
-   const clear = useClearFilter();
-   const valuesById = state.facetValues.byId;
-
-   const value = React.useMemo(() => {
-      const facetName = formatFacetName(filter.facetName);
-      switch (filter.type) {
-         case 'basic': {
-            return `${facetName}: ${valuesById[filter.valueId].value}`;
-         }
-         case 'numeric-comparison': {
-            return `${facetName}: ${filter.operator} ${filter.value}`;
-         }
-         case 'numeric-range': {
-            return `${facetName}: ${filter.range.min} - ${filter.range.max}`;
-         }
-         default:
-            return assertNever(filter);
-      }
-   }, [filter, valuesById]);
-
-   const clearFilter = React.useCallback(() => {
-      clear(filter.id);
-   }, [clear, filter.id]);
-
-   return (
-      <Tag size="md" variant="outline" colorScheme="brand">
-         <TagLabel>{value}</TagLabel>
-         <TagCloseButton onClick={clearFilter} />
-      </Tag>
    );
 };
