+++ features/collection/components/SubcategoriesSection.tsx
@@ -1,7 +1,6 @@
 import {
    Box,
    Button,
-   Center,
    Collapse,
    Flex,
    Heading,
